Move button to header
